+++ docs/pages/release_notes.rst
@@ -12,6 +12,8 @@ Release date: ?
 - Some VDR OpenAPI operations specified ``application/json+did-document`` as Content-Type, while they actually returned ``application/json``.
   This inconsistency is fixed by changing the OpenAPI specification to ``application/json``.
 - Diagnostics now show the conflicted document count for DID Documents the node controls. See monitoring documentation for more detail.
+- ``network.connections.outbound_connectors`` on ``/status/diagnostics`` has been moved to ``/internal/v1/network/addressbook``.
+  Previously it showed only failing connections, now it shows all addresses it will try to connect to (regardless it's already connected to them or not).
 
 
 *************************